refactor: move mineGameLogic to `services`
