+++ services/mineGameLogic.ts
